backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
httpx<0.24.0,>=0.23.0
supabase==1.0.3
oandapyV20==0.7.2
pandas<2.0.0,>=1.3.5
pandas-ta==0.3.14b
numpy==1.25.2
scikit-learn==1.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.8.5
plotly==5.16.1
newsapi-python==0.2.7
pytz==2023.3
finbert-embedding==0.1.5
transformers==4.33.1
torch==2.0.1
openai==1.3.5
alpaca-py==0.8.2
websockets==11.0.3
jwt==1.3.1
celery==5.3.4
redis==5.0.0
